Gate output control signals to memory from LB controller on fsm_en
Now, sram's will only be enabled (in LB mode) when fsm_en (which is the
top level wen_in signal) is high. This ensures that when fsm_en is low,
"everything is paused". We achieve this by using the LB control outputs
to set chip enable to low for both srams.